requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.77.0
openpyxl==3.1.5
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
PyMuPDF==1.25.5
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
tkinterdnd2==0.4.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
